Extract command and event logic, cleanup
